fix(dbt): cast legende fields as float

+++ dbt/models/noisemap/marts/noisemap.sql
@@ -17,7 +17,7 @@ SELECT
   typesource,
   cbstype,
   zonedef,
-  CAST(legende as int) AS legende,
+  CAST(legende as float) AS legende,
   indicetype,
   validedeb,
   validefin,
